requirements.txt
# Run the following command:  pip install -r requirements.txt
Flask==3.0.3
flask-mysqldb==2.0.0
google-generativeai==0.8.3
beautifulsoup4==4.12.3
requests==2.32.3
pytrends==4.9.2
mysqlclient==2.2.5
python-dotenv==1.0.1
scikit-learn==1.5.2
werkzeug==3.0.6
